Moving `Ranges` to other package
