+++ src/main/java/usi/si/seart/converter/SearchParameterDtoToGitRepoSearchConverter.java
@@ -3,9 +3,9 @@ package usi.si.seart.converter;
 import lombok.AllArgsConstructor;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.lang.NonNull;
+import usi.si.seart.collection.Ranges;
 import usi.si.seart.dto.SearchParameterDto;
 import usi.si.seart.repository.specification.GitRepoSearch;
-import usi.si.seart.util.Ranges;
 
 @AllArgsConstructor
 public class SearchParameterDtoToGitRepoSearchConverter implements Converter<SearchParameterDto, GitRepoSearch> {
